Spambam/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.3.0
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
pytest==8.4.1
httpx==0.28.1
requests==2.32.4